Simplify `resolve_ws` API
