+++ src/cargo/ops/cargo_fetch.rs
@@ -1,22 +1,12 @@
-use core::registry::PackageRegistry;
-use core::{PackageId, Resolve, PackageSet, Workspace};
+use core::{Resolve, PackageSet, Workspace};
 use ops;
 use util::CargoResult;
 
 /// Executes `cargo fetch`.
 pub fn fetch<'a>(ws: &Workspace<'a>) -> CargoResult<(Resolve, PackageSet<'a>)> {
-    let mut registry = PackageRegistry::new(ws.config())?;
-    let resolve = ops::resolve_ws(&mut registry, ws)?;
-    let packages = get_resolved_packages(&resolve, registry);
+    let (packages, resolve) = ops::resolve_ws(ws)?;
     for id in resolve.iter() {
         packages.get(id)?;
     }
     Ok((resolve, packages))
-}
-
-pub fn get_resolved_packages<'a>(resolve: &Resolve,
-                                 registry: PackageRegistry<'a>)
-                                 -> PackageSet<'a> {
-    let ids: Vec<PackageId> = resolve.iter().cloned().collect();
-    registry.get(&ids)
 }
